[web] Add logic for displaying only chats from a given community in calendar filters
Summary:
issue: https://linear.app/comm/issue/ENG-2794/make-it-possible-to-display-only-chats-from-a-given-community-in-the
Add to `FilterPanel` logic, that will allow to display only chats of a given community in filters. For now `onSetCommunity` and `onResetCommunity` are not called from anywhere.

Test Plan:
Called `onSetCommunity`, checked that calendar filters changed accordingly, and so did events displayed in the calendar. Checked that searching works properly - displaying only chats from the
chosen community. Checked that the number of
threads displayed under "Your chats" is correct (also when searching - that it is reduced to the number of matches from the chosen community).
Checked that `onResetCommunity` works - makes all filters available.
Checked that when a community is chosen it is not possible to add an event for a chat from a different community - ThreadPickedModal displayes only chats from this community.

Reviewers:
kamil, tomek
Subscribers: