[keyserver] Add thread column to uploads table
Summary:
Attaching a thread column to the uploads table will important so
we can associate each upload with a thread. This will allow us to later
query media that was uploaded in a specific thread.

Test Plan:
Restart `keyserver` and check the tables on TablePlus for
`uploads` to check and confirm that there is a new `thread` column.

Reviewers: atul, ginsu, kamil, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6465