[keyserver] Add thread column to uploads table
Summary:
Attaching a thread column to the uploads table will important so we can associate each upload with a thread. This will allow us to later query media that was uploaded in a specific thread.

https://linear.app/comm/issue/ENG-2862/modify-the-uploads-table-to-include-a-thread-column (uploads table modification)

https://linear.app/comm/issue/ENG-2883/create-a-script-to-retroactively-populate-the-thread-column-in-the (retroactively populate table)

Test Plan:
Restart `keyserver` and check the tables on TablePlus for `uploads` to check and confirm that there is a new `thread` column.

{F352255}

Reviewers: atul, ginsu, kamil, tomek, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6465