[native] Create a component for the media gallery on native
Summary: Here, we create the native component for the thread media
gallery. Currently, `mediaInfos` is defaulted to an empty array, but it
will be replaced to hold the media data in subsequent diffs.
`galleryItemGap` refers to the gap between media, and
`galleryFirstItemGap` refers to the first media in each row.
`galleryItemWidth` is determined using `Dimensions` so we can hopefully
ensure that each item takes 1/3 of the width of the gallery.

Encompassing the media in <TouchableOpacity> will be changed in later
diffs, alongside just rendering an <Image> component.

Depends on D6465

Test Plan: This will be tested in the next diff, and a screenshot of what
the media gallery will look like with sample, hard coded images will be
shown.

Reviewers:

Subscribers: