[lib] Inline threadSearchText
Summary:
This function is only used from one place in the codebase. I want to separate it out into two parts – one for `UserInfo` fetching and another for the rest. That's because I want to pass the `UserInfo`s through `useENSNames`.

Depends on D6478

Test Plan: Flow, and tested with the rest of the stack

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6479