[web] Use app switcher in the topbar
Summary:
issue: https://linear.app/comm/issue/ENG-2739/move-navigationpanel-from-the-sidebar-to-a-top-bar
Use `AppSwitcher` in `Topbar`. Now, that the `AppSwitcher` is inside of the `Topbar`, it can be made horizontal, and it can use the styles for a horizontal switcher. The styles for icons in the `left-layout-aside.css` are no longer needed,
because the other switcher doesn't display any icons.

Test Plan: use `Topbar` in `App` component, check that it displayes correctly.

Reviewers: przemek, michal, bartek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6345