[keyserver] Create a function to query for all uploads from a provided threadID
Summary: We create a function in `upload-fetchers` to query the database for uploads given a specific threadID. The `filename NOT LIKE 'thumb%'` is used to filter out scenarios where when a video is uploaded,
so is a thumbnail, and we don't want to pull a static thumbnail for the video.

Depends on D6482

Test Plan: I tested this locally in `keyserver.js`, by importing the function and running this code:

```
fetchMediaForThread("85372").then((media) => {
console.log(media);
});
```

Checking the console logs confirms that the media is successfully fetched.

Reviewers: atul, ginsu, kamil, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6485