[native] introduce showEmojiKeyboard shared value
Summary:
introduce `showEmojiKeyboard` shared value. This will be used to open/close the emoji keyboard and open/close the tooltip, which will be implemented in a subsequent diff.

Also noticed as developing that `setHideTooltip` would be better typed as `SetState<boolean>` so included that change in this diff.

---

Linear Task: [[ https://linear.app/comm/issue/ENG-2777/implement-native-emoji-keyboard-library | ENG-2777 ]]

Test Plan: `flow` and diff will be further tested in subsequent diffs

Reviewers: atul, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6445