[services] Tunnelbroker - Expose `isConfigParameterSet` function to the Rust
Summary:
This function exposes the `isConfigParameterSet` function from the C++ codebase to the Rust by using the CXX function.
This function will be used in the next D6253 diff to retrieve the boolean status if the config file flag.

Linear task: [[ https://linear.app/comm/issue/ENG-2642/disabling-authentication-mechanism-by-the-config-flag-in-tunnelbroker | ENG-2642 ]]

Test Plan: Adding the `skip_authentication` flag to the `[sessions]` section of the config file results in a `true` return from the `isConfigParameterSet('sessions.skip_authentication')` function call from the Rust codebase.

Reviewers: jon, bartek

Reviewed By: jon, bartek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6252