[lib] Add messageInfoType field to ChatMessageInfoItem
Summary:
In a future diff I'm going to be changing the type of `robotext`, so our current `typeof robotext === 'string'` runtime assertion won't be as easy to make.

Instead, I extend the types in this diff to make a different kind of easy runtime assertion possible.

Depends on D6505

Test Plan: Flow

Reviewers: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6506