[web] Pull `SIWEButton` out of `SIWE` component
Summary:
We introduce `SIWEButton` and continue to use as-is in `SIWE` component.

In subsequent diff we'll move `SIWEButton` out of `SIWE` into `LoginForm` so it can be used independently (just to initiate the SIWE auth flow).

---

Depends on D6488

Test Plan: This diff is a noop, tested the SIWE web flow and things continue to work as they did before.

Reviewers: ashoat, tomek, ginsu, rohan

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6489