[lib] Use term "channel" in threadNoun
Summary:
This diff updates our copy to use the term "channel" when referring to a chat inside a community. This will be used in robotext and notifs.

Depends on D6528

Test Plan: Flow and careful reading through the code. `threadNoun` is only used in two places – once in `getNameForThreadEntity` in `entity-text.js`, and once in `notifTextForSubthreadCreation`. I made sure that the relevant robotexts that end up using `threadNoun` through `ET` were updated

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6529