[lib] Simplify messageTitle and introduce default for robotext
Summary: This diff switches `messageTitle` to returning `EntityText` as preparation for ENS resolution. I take this opportunity to simplify the logic for `global_viewer`, and to delete most of the implementations of `messageTitle` in favor of a default for robotext.

Test Plan: Tested in combination with following diffs, where the message title is actually used. Also Flow

Reviewers: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6560