[keyserver] Fix publishing new messages / latest messages when a new message is created.
Summary:
In D6224 we've introduced a bug that causes new messages to not always be delivered to some users. The issue is described in https://linear.app/comm/issue/ENG-2889/threads-arent-marked-as-unread-on-web.

This solution modifies the logic so that we only skip the part responsible for notifications when device list is empty.

Test Plan: Open two web clients with different users logged in. Make it so one of them doesn't have device id set in any of the cookies. Check if before this diff only one user receives new messages / thread unread status updates. Check if after this diff both users receive these.

Reviewers: kamil, inka, ginsu, atul, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6530