[web] Introduce `OrBreak` component
Summary:
We pull the "or" `<hr />` into a separate component so we can continue using it across `LoginForm` and `SIWELoginForm` while also being able to use a non-"or" `hr` within `SIWELoginForm`.

Doing this with CSS was pretty messy, pulling this into it's own component made it cleaner/simpler.

Test Plan: Looks as expected. The existing "or" breaks look as expected, and non-"or" `hr`s also look as expected.

Reviewers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D6535