[lib] Return EntityText from getMessageTitle
Summary:
We need to resolve this `EntityText` at callsites, so we'll return it directly.

Initially the callsites will call `entityTextToRawString` same as before, but in following diffs I'll refactor it.

Depends on D6561

Test Plan: Tested in combination with following diffs, where the message title is actually used. Also Flow

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6562