[lib] Catch exceptions thrown by ethers.js in ENSCache
Summary:
I haven't noticed any exceptions, but it would be good to consider this.

If an exception is thrown when querying for an address, we won't try again for another 5 minutes.

Depends on D6550

Test Plan: I haven't actually noticed exceptions, but in a later point of the stack I introduce my own exceptions (for timeouts) and I confirmed things work correctly for those

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6551