[lib][web][native] Fetch ENS names for sidebar preview
Summary:
With this diff, message titles are now handled by `EntityText`, and the `EntityText` is fully resolved at all relevant callsites.

Depends on D6566

Test Plan:
Add artificial delays to ENS fetching and confirm delays reflected in sidebar creation, and that sidebar title in UI includes resolved ENS names.

Note that there are two places in `native` that I decided not to fetch ENS names in: `SidebarInputBarHeightMeasurer` and `useAnimatedMessageTooltipButton`. I checked and both of these places don't look at the thread title. The former handles height measurement and the latter handles animations. In both cases a hook would be inappropriate as it would cause the process (height measurement or animation setup) to have to be done twice (once after ENS resolution), which would be would be performance/efficiency with no benefit. Using an `async` function here would avoid forcing things to happen twice, but it would unnecessarily delay things and make performance worse.

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6567