[lib] Migrate CreateThread/CreateSubThread notifTexts to use EntityText
Summary:
Note that this introduces a require cycle by having message specs import directly from `notif-utils`, which imports into them. I think it's fine in modern Node.js.

Depends on D6577

Test Plan: Generated `CreateThread` and `CreateSubThread` messages on a web client in a chat with a different test user that is logged in on the iOS simulator. Background the iOS app and make sure the notif looks right

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6578