[lib] Update ThreadInfo.uiName to return an array of UserEntities
Summary:
This diff enables all of the diffs below it in the stack to actually work. I had to wait until the last diff to turn everything on because it otherwise would break Flow types.

Depends on D6614

Test Plan: Tested in combination with all of the diffs earlier in the stack

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6615