[web] Introduce basic `ConnectedWalletInfo` component
Summary: This component will replace the the Rainbowkit `ConnectButton`. For now just displays the ENS name or address of connected wallet. In future diffs will also display avatar (ENS or rainbowkit emoji) and will display `AccountModal` when clicked.

Test Plan:
Looks as expected:

{F360510}

{F360511}

> Should we shorten the address?

Yeah I'll address that in subsequent diff.

> Why are we keeping the `ConnectButton` around?

It's been helpful when working on this stack to be able to have them side-by-side. Since all of this is hidden behind an `isDev` check it just makes it more convenient instead of stashing/popping over and over again.

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6570