[web] Use SettingsSwitcher and Topbar
Summary:
issue: https://linear.app/comm/issue/ENG-2739/move-navigationpanel-from-the-sidebar-to-a-top-bar
The `AppSwitcher` is now displayed in the topbar. The `SettingsSwitcher` is displayed as it used to, but I moved it out of the sidebar because of how we use the grid.
Our grid: https://github.com/CommE2E/comm/blob/master/web/style.css#L54
So the SettingsSwitcher has to be a part of app grid area, or the grid layout has to change.
Changing the way we use the grid - or not using the grid - can be done, but it seems like a really big side quest to my community drawer task. So for now I think this is a good enough solution.
It is also consistent between the `AppSwitcher` and `SettingsSwitcher`, so I think it's not necesserly bad.

Test Plan: Run web app, check that the `SettingSwitcher` and `AppSwitcher` display and work correctly.

Reviewers: przemek, michal, bartek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6346