[lib] Pluralize NotifTexts.title when it's not a string
Summary:
In an intermediate revision of [D6586](https://phab.comm.dev/D6586) I switched `ThreadInfo.uiName` from returning a pluralized `EntityText` directly, to returning an array of `UserEntity`s that still need to be pluralized. I did this because in some cases we'd want to turn off the behavior of dropping the viewer from the list of viewers, and we would need to do that before pluralization.

However, I forgot to update the code for determining notif texts. Flow didn't catch this because technically an array of `UserEntity`s **is** an `EntityText`, so it didn't see any issues. But the result of this is that the pluralization was skipped, and we got the behavior reported in [ENG-2932](https://linear.app/comm/issue/ENG-2932/notification-title-displaying-without-spaces), which I hadn't noticed during testing since I only tested iOS.

Test Plan:
Tested notif in Android emulator and make sure title appears properly pluralized

| {F366718} |

Reviewers: inka, tomek, atul

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D6627