[lib] Show "just you" for thread title if there are no members besides you
Summary:
This resolves [ENG-2933](https://linear.app/comm/issue/ENG-2933/calendar-filters-displaying-incorrectly-on-master), which was broken in D6615 when I removed the `'just you'` special-case there.

Since filtering out the user (and pluralization) are now handled in `getNameForThreadEntity`, we need to check the length of the list there instead of in `threadUIName` now.

Test Plan: Make sure there are no empty strings in the web calendar filters, and that 'just you' appears where the empty strings previously did

Reviewers: inka, tomek, atul

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D6628