[web] Make pressing a drawer item in calendar tab change available filters
Summary:
issue: https://linear.app/comm/issue/ENG-2793/make-calendar-filters-change-on-drawer-action
We want pressing on a community drawer item, in the calnedar tab, to result in only events from chats from that community being displayed, and additionally only chats from this community being present in
the filter panel.
`CalendarDrawerItemHandler` is a component, that provides a `handler` object, that has the `onClick` function for the drawer items for the calendar tab. This diff channges what this 'onClick' function
does, to satisfy the requirement from above.

Test Plan:
Run web app, checked that pressing on an item in the community drawer results in only events from chats from that community being displayed, and only chats from this community being present
in the filter panel.

Reviewers:
tomek, kamil

Subscribers: