[web] Add `shortenAddressToFitWidth` to `ConnectedWalletInfo`
Summary:
We want to shorten addresses and long ENS names to fit within `ConnectedWalletInfo`. Similar to the RainbowKit `ConnectButton`, we add an ellipsis in the middle of the string if it's too long and display 10 characters of the prefix and suffix. We show the full address on hover so the user can check the full address (which RainbowKit doesn't do, but figured it doesn't hurt).

---

Depends on D6633

Test Plan:
Here's what it looks like in various scenarios:

{F367239}

{F367240}

{F367241}

{F367245}

Reviewers: ashoat, tomek, ginsu, rohan

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6635