[yarn] Introduce `yarn ci-cleaninstall`
Summary:
Move the `cleaninstall` command with whatever flags used on CI to `package.json` to reduce length of command in `eslint_flow_jest.yml` and to be able to use throughout the repo with consistent flags (eg if we want to change timeout we don't need to do it 100 places).

(This is the last diff in the stack)

Test Plan: Tested locally, trusting this will work... will revert if there are issues.

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6651