[native] Introduce setRouteParams action on OverlayRouter
Summary:
`setParams` lets us set params for the current route, but we need to set params for another route.

`navigate` lets us set params for an existing route, but because it doesn't know if the route already exists or not, it needs the full list of params (instead of just a `$Shape` of them).

This new action will only work if the route exists already.

Test Plan:
1. Opened robotext and text tooltips
2. Mocked an extra `TooltipItem` to show action sheet
3. Made sure tooltip is hidden when action sheet is opened
4. Selected an emoji from the popover. Confirmed everything dismisses at the same time
5. Selected an emoji from the emoji picker. Confirmed everything dismisses at the same time
6. Pressed the background and confirmed everything dismisses at the same time
7. Pressed the copy button to make sure that everything dismisses at the same time

Reviewers: ginsu, tomek, atul

Reviewed By: ginsu

Differential Revision: https://phab.comm.dev/D6657