[native] Clean up Tooltip transition state
Summary:
Currently we have three places state about Tooltip transition is stored:

1. `actionSheetShown`
2. `showEmojiKeyboard`
3. `hideTooltip`

This diff moves `hideTooltip` to ReactNav params so it can be modified outside of the component. The `showEmojiKeyboard` stuff has been moved to the individual button components. And `actionSheetShown` is no longer necessary, and has been deduped with the others.

Depends on D6657

Test Plan:
1. Opened robotext and text tooltips
2. Mocked an extra `TooltipItem` to show action sheet
3. Made sure tooltip is hidden when action sheet is opened
4. Selected an emoji from the popover. Confirmed everything dismisses at the same time
5. Selected an emoji from the emoji picker. Confirmed everything dismisses at the same time
6. Pressed the background and confirmed everything dismisses at the same time
7. Pressed the copy button to make sure that everything dismisses at the same time

Reviewers: ginsu, tomek, atul

Reviewed By: ginsu

Differential Revision: https://phab.comm.dev/D6658