[web] Add calendar filter type for displaying only one community
Summary:
issue: https://linear.app/comm/issue/ENG-2794/make-it-possible-to-display-only-chats-from-a-given-community-in-the
We want to be able to press some button, and see only chats from a given community to be visible in the calendar filters. We want this information to persist when switching between tabs. So we will keep
this as a calendar filter in redux.
This diff introduces the architectural changes required for this - adding and changing types according to what will be needed.

Test Plan: Run web app, check that calendar filters work like they used to.

Reviewers: kamil, tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6457