[keyserver] Change device_token to mediumtext
Summary:
Part of [EMG-2825](https://linear.app/comm/issue/ENG-2825/save-subscription-information-on-the-keyserver).

We need to save web push subscription info on the keyserver inside the `device_token`. Unfortunately it's too long for the current `VARCHAR(255)` so this diff changes that to a `mediumtext`. Additionaly indices had to be modified because of the the type change. Now we need to specify a prefix length that they apply to. I've tested it and it seems like registration info is 400-500 bytes so I've set it to 512.

Test Plan: Check if migration is correctly applied.

Reviewers: atul, kamil, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D6714