[keyserver/web] Pass push notif API public key to web app
Summary:
[ENG-2825](https://linear.app/comm/issue/ENG-2825)
Keyserver needs to authenticate with the push service and to do that we need to generate a key pair and send the public one to the web app, which it will use during notification subscription. The keys were generated using `web-push` package in a json format. This diff injects the push key into redux.

Test Plan: Check if the key is available in redux in the web app.

Reviewers: inka, atul, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6717