[web] Highlight drawer items in calendar tab
Summary:
issue: https://linear.app/comm/issue/ENG-2839/highlight-drawer-item-in-the-calendar-tab
We want the drawer item corresponding to the picked communnity to be highlithed (in calendar tab). `isActive` property of `handled` object is what defines that behaviour - drawer item based on it
appleis the highlight or not.

Test Plan:
Run web app, checked that drawer items highlight propelry, and when no comunity is picked, no item is highlighted.

Reviewers:
kamil, michal

Subscribers: