[web] Add All communities / Inbox button
Summary:
Issue: https://linear.app/comm/issue/ENG-2887/create-all-communities-button
As discussed in ENG-1882 we want to have a "All communities" button, that resets the calendar filters to show all chats the user is a member of. This is in contrast to ENG-2794, where we create the
possibility to display in filters panel only chats from a given community.
The styles will be revisited during https://linear.app/comm/issue/ENG-2671/style-community-drawer-on-web

Test Plan:
Run web app, check that at the top of the drawer:
1. in calendar tab there is a "All communities" button, and pressing it resets the community filter
2. in all other tabs there is a "Inbox" button, and pressing it navigates back to chat tab

Reviewers:
michal, przemek

Subscribers: